create brain/selectivity; create new test file for augmented_nn.
