+++ src/brain/workload/workload_defaults.cpp
@@ -36,15 +36,5 @@ const int LinearRegWorkloadDefaults::BPTT = 90;
 
 const int KernelRegWorkloadDefaults::BPTT = 90;
 
-
-const int AugmentedNNWorkloadDefaults::COLUMN_NUM = 1;
-const int AugmentedNNWorkloadDefaults::ORDER = 1;
-const int AugmentedNNWorkloadDefaults::NEURON_NUM = 16;
-const float AugmentedNNWorkloadDefaults::LR = 0.1f;
-const int AugmentedNNWorkloadDefaults::BATCH_SIZE = 256;
-const int AugmentedNNWorkloadDefaults::EPOCHS = 600;
-
-
-
 }  // namespace brain
 }  // namespace peloton
